requirements.txt
cachetools==5.5.2
certifi==2025.4.26
cryptography==44.0.3
dnspython==2.7.0
google-ads-admanager==0.8.0
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleads==49.0.0
googleapis-common-protos==1.73.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
isodate==0.7.2
lxml==5.4.0
oauth2client==4.1.3
oauthlib==3.2.2
pymongo==4.12.1
python-dotenv==1.1.0
redis==5.3.0
requests==2.32.3
tqdm==4.67.1
urllib3==2.4.0
zeep==4.3.1
